503: Process the q query string for photos
Previously the 'q' query string was not processed and the server would
return all photo albums. Now a filter is applied as expected and only
matching albums are returned. The search box on the website will only
process a request once more than 2 characters are entered to prevent
excessive traffic to and from the server. Since the query string is not
automatically sent this restriction has not been applied on the query
string.