templates: Use IA-specific books URL for everyone
Wo4you changed their store so anyone with a personal link will get a popup saying that people need to visit via the association website. But it is unclear that people need to be logged in to see the correct link. Because Wo4you also asks everyone if they are a member for the discount, we decided that the personal link is OK to be default for everyone.