Fix AttributeError 'zoneinfo.ZoneInfo' object has no attribute 'localize'
Pytz is being deprecated and since Django 4.0 it now returns zoneinfo objects instead of pytz objects. This fix replaces the usage of .localize (which exists on pytz objects) with a .replace call (which is the way to do the same action on regular datetime objects now).

Docs Reference: https://docs.djangoproject.com/en/5.1/releases/4.0/#zoneinfo-default-timezone-implementation

Fixes crashes on personal tab transactions pages, as well as anywhere else in the code where the .localize function was used.